Convertento dados de entrada 2

+++ api.js
@@ -40,7 +40,7 @@ app.post("/sensor", async (req, res) => {
 
   try {
     const newReading = await prisma.silo.update({
-      where: { id: parseFloat(id) },
+      where: { id: parseInt(id) },
       data: {
       sensor_value: parseFloat(sensor_value),
       dt_register: new Date(),
